refactor(mobile): split channels page widgets
